feat, support upgrade for paras
